backend/requirements.txt
torch
clip @ git+https://github.com/openai/CLIP.git
faiss-cpu
numpy
Pillow
flask
flask-cors
